Uploading full script to initialize the repo for the first time
The project was developed as a means to create a prototype combat
engine for an RPG being developed for the Sega Genesis (a retro
video game console).

The end product is intended to be translated over to a development
kit designed specifically for the Sega Genesis. For now, the
project lives as a text-adventure.

Plans for the future include:

Split up scripts to increase readability

An enemy AI that attacks back

A menu system

Possibly a walking simulation

Ideally the game will be ported over to the Sega Genesis so visual
assets can be added

A game over screen