Reorganized header files and started first state
Found an appropriate home for the call stack components. Initially
it was being init. on multiple different .c files which was causing
duplicate errors when everything was compiled. Instead I created
a messages.h and messages.c files where all of the games messages
will live.

I also started work on our first state, the exploration state.
Next is to start working in conditions that'll trigger other states
and help move along the flow of gameplay.