Squashed bugs in battle and exploration files
Previously there was a bug that would occur during the init. battle
function that caused a hit to not connect.

Was also able to get the main block to transition b/t two different
states. There are also states that switch on the battle / exploration
functions, however this needs to be altered to fit a more modular
state machine. The state machine should be operating in the main block
not on individual functions.

I was also able to create a funciton that creates Entities to be
used dynamically in other contexts.