Fixed bug causing errors when creating entities
Previously when creating new entities for a context (i.e. a battle
or map) the entities stats were not properly being initialized.
I found an error in the switch statements used to
determine which entity is being created where break statements
were not included. Adding break statements resolved the problem.