Setup a better implementation of the state machine
To make it easier to have certain contexts trigger other states,
I made the state functions that invoke certain contexts, return a
state that can shange what functions are invoked in the main block.

This allows us to define the functional logic in our seperate C
files, but control the flow of the gameplay in the main block.

The next step is to break up the initExploration and initBattle
functions into smaller contexts. That way we can have states
control a more modular set of functions.