Implemented nested states to channel flow through exploration state.
Broke out the exploration state into multiple sub-states so code
is clearer and more modular. In particular, initExploration was
broken up into a variety of sub-states that move the player,
check if an enemy is present, and if so, change the parent state
from Exploration to Battle.

Next step is to allow enemies to move on patrol and break out the
Battle state into multiple sub states.