Added map bounds and refactored exploration state
Added boundaries on all sides of the map as well as conditional
logic to prevent entities from going off the map.

I also expanded the size of the 4 x 4 map to 5 x 5. Furthermore,
I simplified code so that the overall exploration state is
more readable for future developers on the project. Finally,
I ensured that static enemies on the map, will trigger an
enemy encounter if the player hits an enemy location.

Next feature is to allow enemies to walk along a pre-defined
path.