Added a circular walk path for the Small Monster
Need to figure out a more efficient way to implement this. Currently
this is not the most scalable way to do things as it requires you
to declare a static location for each monster. If we can figure out
a way to implement a closure like function so that monsters can
walk in a unique pattern for each, that would be ideal.

I also need to clean up the code to be more readable.