Build can now trigger a battle state with specific enemies
Prior the battle state was triggered when the player collides with
any enemy and the battle would be waged with the same enemy
regardless of who the player encounters. Now the build is able to
determine which enemy the player collides with and passes the
enemy entity into the battle state so the player can fight against
the specific enemy they triggered.

I also refactored code in the exploration state to be more concise
and utilize less memory. I also removed any states that are unused.