Successfully transcribed entity.c to SGDK compliant code
I was able to successfully transition the original vanilla C code
for entity.c over to SGDK compliant code with memory allocation
and safe code. There should be no instances of memory leakage caused by
depreciated functions or strings.

Next step is to do the same for input so the user can enter commands
and the game will register input.