Added sprite animations
In addition to importing in a new sprite sheet, I setup the built-in
funcs necessary for starting animation cycles. We also created an easy
format to plug in our animation cycles from imported sprite sheets.

Next step is to abstract everything into it's own set header, c file
and functions so that we can invoke them anywhere through out the game
depending on the state they're in.