Abstracted Sprites header and init. walking animation
After abstracting the sprites function into their own header and
was able to implement sprites into the exploration state. We might
need to be careful of declaring an empty pointer in the exploration
state. We have to see where we declare the sprites pointer and
where we init. them.

For testing purposes, we're declaring and init. pointers in each
state. We also need to get a different sprite sheet in motion
as the current one is for a beat-em-ups and less for an RPG or
action-adventure.