Added a map background for sprite to walk around on
Initialized the game to have a map with a background from the
Stellar Winds assets. Currently the sprite can walk around the
map, but there is no collission and the map doesn't move
around with the player as they move.

The next step is to start the sprite in an area where there
is no colission events and have the map scroll with the player
gradually.